Several changes to the SFX lock code in the sound engine, mainly to make sure SFXs are automatically freed only at level change Moved call to CL_InitTEnts after S_Startup so the tent sounds get properly precached Added a developer warning when trying to play a non-precached SFX Made the soundlist command more verbose Added the ssize_t type for Win32 systems Some comments and dead code removal
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@4781 d7cf8633-e32d-0410-b094-e92efae38249

+++ fs.c
@@ -405,10 +405,10 @@ qboolean PK3_GetEndOfCentralDir (const char *packfile, FILE *packhandle, pk3_end
 	buffer = Mem_Alloc (tempmempool, maxsize);
 #ifdef FS_USESYSCALLS
 	lseek (packhandle, filesize - maxsize, SEEK_SET);
-	if (read (packhandle, buffer, maxsize) != (int) maxsize)
+	if (read (packhandle, buffer, maxsize) != (ssize_t) maxsize)
 #else
 	fseek (packhandle, filesize - maxsize, SEEK_SET);
-	if (fread (buffer, 1, maxsize, packhandle) != (unsigned int) maxsize)
+	if (fread (buffer, 1, maxsize, packhandle) != (size_t) maxsize)
 #endif
 	{
 		Mem_Free (buffer);
